Move ValueObjects inside EventStore namespace
