+++ src/EventStore/ValueObjects/Identity/UUID.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace ValueObjects\Identity;
+namespace EventStore\ValueObjects\Identity;
 
-use ValueObjects\Exception\InvalidNativeArgumentException;
-use ValueObjects\StringLiteral\StringLiteral;
-use ValueObjects\Util\Util;
-use ValueObjects\ValueObjectInterface;
+use EventStore\ValueObjects\Exception\InvalidNativeArgumentException;
+use EventStore\ValueObjects\StringLiteral\StringLiteral;
+use EventStore\ValueObjects\Util\Util;
+use EventStore\ValueObjects\ValueObjectInterface;
 use Rhumsaa\Uuid\Uuid as BaseUuid;
 
 class UUID extends StringLiteral
