+++ tests/EventStore/ValueObjects/Tests/TestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ValueObjects\Tests;
+namespace EventStore\ValueObjects\Tests;
 
 class TestCase extends \PHPUnit_Framework_TestCase
 {
